Refactor player state transitions
Updated state handling in Player and StateMachine classes, removing redundant move check in InputManager and improving transition logic at the end of various states. This makes the player character's movement more fluid and reduces duplication of logic in the code. Specifically transition statements were added to FallState and AttackState classes and removed from MoveState class. A check for movement input was added whenever a state transition to Idle was likely to happen, this is to ensure that the player transitions to Walk state instead of Idle if they're still holding a movement key.