Add new component and modify input settings
A new component was added to the Player Input Manager in the Sandbox scene, and the default control scheme was changed to 'Gamepad'. This was necessary to add more functionality to the player and improve gameplay experience for gamepad users. Also, in MoveState.cs, a comment line was edited to properly reflect the behaviour when transitioning to Idle state. This modification helps to prevent the player from gaining too much speed when movement starts after being idle.