Remove Timer.prefab and update input control and RoundTimer.cs
Removed an unnecessary Timer.prefab.
Updated the Player Input Manager from using rightShift to tab for joining action.
If the RoundTimer is not custom, it's preferred to use the InvariantCulture instead of CurrentCulture for better uniformity.
Setting CurrentTime to public allows RoundTimerEditor to access and modify its value.
On a side note, EnterPlayModeOptionsEnabled is disabled for debugging purposes.
New Gamepad controls for Jump, Crouch, Modifier and Attack are added for better user experience.
Round Timer is re-introduced in the Development scene.