Add updated movement.
Had to modify to some extent to make it work with the current system. Most notably, Player Rigidbody Mass was moved from the fall state to the jump state, as there would be a bug where the player would jump higher the first time they jumped, as the mass had not been set to it's proper value until after fall state was entered.

In the future, I'd say it would be beneficial to move rigidbody logic to its own space or something in that nature, for instance moving it to the Idle state as every player will always be in the idle state at the start of the game, or even the None state as that also occurs at the start of the game.
This would mean that iteration would become slightly more tedious as you would have to tab into the NoneStateData or IdleStateData to change the mass value, but it could be beneficial towards the overall project.

Signed-off-by: Lumina <alexander.andrejeff@elev.ga.lbs.se>